backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-google-genai
langchain-community
langgraph
pinecone
pypdf
python-dotenv
tiktoken
langchain-experimental
